# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BuildPro Construction Business Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /clients/{id}/history:
    get:
      operationId: getClientHistory
      tags: [clients]
      summary: Get client job history and totals
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientHistory"

  # ──────────────────────────────────────
  # JOBS
  # ──────────────────────────────────────
...
  /estimates:
    get:
      operationId: listEstimates
      tags: [estimates]
      summary: List all estimates
      parameters:
        - name: jobId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of estimates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Estimate"
    post:
      operationId: createEstimate
      tags: [estimates]
      summary: Create a new estimate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateInput"
      responses:
        "201":
          description: Created estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Estimate"
  /estimates/{id}:
    get:
      operationId: getEstimate
      tags: [estimates]
      summary: Get an estimate by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Estimate"
        "404":
          description: Not found
    patch:
      operationId: updateEstimate
      tags: [estimates]
      summary: Update an estimate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateUpdate"
      responses:
        "200":
          description: Updated estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Estimate"
    delete:
      operationId: deleteEstimate
      tags: [estimates]
      summary: Delete an estimate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /estimates/{id}/send:
    post:
      operationId: sendEstimate
      tags: [estimates]
      summary: Mark estimate as sent to client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendEstimateInput"
      responses:
        "200":
          description: Estimate sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Estimate"

  # ──────────────────────────────────────
  # ESTIMATE LINE ITEMS
  # ──────────────────────────────────────
  /estimates/{estimateId}/items:
    get:
      operationId: listEstimateItems
      tags: [estimate-items]
      summary: List all line items for an estimate
      parameters:
        - name: estimateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of estimate line items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EstimateItem"
    post:
      operationId: createEstimateItem
      tags: [estimate-items]
      summary: Add a line item to an estimate
      parameters:
        - name: estimateId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateItemInput"
      responses:
        "201":
          description: Created line item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstimateItem"
  /estimates/{estimateId}/items/{itemId}:
    patch:
      operationId: updateEstimateItem
      tags: [estimate-items]
      summary: Update a line item
      parameters:
        - name: estimateId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateItemUpdate"
      responses:
        "200":
          description: Updated line item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstimateItem"
    delete:
      operationId: deleteEstimateItem
      tags: [estimate-items]
      summary: Delete a line item
      parameters:
        - name: estimateId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ──────────────────────────────────────
  # MATERIALS
  # ──────────────────────────────────────
...
components:
  schemas:
...
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        zipCode:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, createdAt]
    ClientInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        notes:
          type: string
      required: [name]
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        notes:
          type: string
    ClientHistory:
      type: object
      properties:
        clientId:
          type: integer
        totalJobs:
          type: integer
        completedJobs:
          type: integer
        totalRevenue:
          type: number
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
      required: [clientId, totalJobs, completedJobs, totalRevenue, jobs]

    # ── JOBS ──
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        status:
          type: string
          enum: [lead, scheduled, in_progress, completed, cancelled]
        description:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        zipCode:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        estimatedValue:
          type: ["number", "null"]
        actualCost:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, title, status, createdAt]
...
    Estimate:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: ["integer", "null"]
        jobTitle:
          type: ["string", "null"]
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        title:
          type: string
        status:
          type: string
          enum: [draft, sent, accepted, rejected, revised]
        subtotal:
          type: number
        taxRate:
          type: number
        taxAmount:
          type: number
        total:
          type: number
        notes:
          type: ["string", "null"]
        sentAt:
          type: ["string", "null"]
        validUntil:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, title, status, subtotal, taxRate, taxAmount, total, createdAt]
    EstimateInput:
      type: object
      properties:
        jobId:
          type: integer
        clientId:
          type: integer
        title:
          type: string
          minLength: 1
        status:
          type: string
          enum: [draft, sent, accepted, rejected, revised]
        taxRate:
          type: number
        notes:
          type: string
        validUntil:
          type: string
      required: [title]
    EstimateUpdate:
      type: object
      properties:
        jobId:
          type: integer
        clientId:
          type: integer
        title:
          type: string
        status:
          type: string
          enum: [draft, sent, accepted, rejected, revised]
        taxRate:
          type: number
        notes:
          type: string
        validUntil:
          type: string
    SendEstimateInput:
      type: object
      properties:
        clientEmail:
          type: string
        message:
          type: string
      required: [clientEmail]

    # ── ESTIMATE ITEMS ──
    EstimateItem:
      type: object
      properties:
        id:
          type: integer
        estimateId:
          type: integer
        description:
          type: string
        category:
          type: string
          enum: [labor, material, equipment, subcontractor, other]
        quantity:
          type: number
        unit:
          type: string
        unitPrice:
          type: number
        totalPrice:
          type: number
        notes:
          type: ["string", "null"]
        sortOrder:
          type: integer
      required: [id, estimateId, description, category, quantity, unit, unitPrice, totalPrice, sortOrder]
    EstimateItemInput:
      type: object
      properties:
        description:
          type: string
          minLength: 1
        category:
          type: string
          enum: [labor, material, equipment, subcontractor, other]
        quantity:
          type: number
        unit:
          type: string
        unitPrice:
          type: number
        notes:
          type: string
        sortOrder:
          type: integer
      required: [description, category, quantity, unit, unitPrice]
    EstimateItemUpdate:
      type: object
      properties:
        description:
          type: string
        category:
          type: string
          enum: [labor, material, equipment, subcontractor, other]
        quantity:
          type: number
        unit:
          type: string
        unitPrice:
          type: number
        notes:
          type: string
        sortOrder:
          type: integer

    # ── MATERIALS ──
...