# lib/api-zod/src/generated/types/estimateItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BuildPro Construction Business Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { EstimateItemInputCategory } from './estimateItemInputCategory';

export interface EstimateItemInput {
  /** @minLength 1 */
  description: string;
  category: EstimateItemInputCategory;
  quantity: number;
  unit: string;
  unitPrice: number;
  notes?: string;
  sortOrder?: number;
}