# lib/api-zod/src/generated/types/estimateItemUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BuildPro Construction Business Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { EstimateItemUpdateCategory } from './estimateItemUpdateCategory';

export interface EstimateItemUpdate {
  description?: string;
  category?: EstimateItemUpdateCategory;
  quantity?: number;
  unit?: string;
  unitPrice?: number;
  notes?: string;
  sortOrder?: number;
}